垃圾清理模块完成

+++ opsxpro/src/main/java/com/zzzmode/appopsx/AssetsUtils.java
@@ -30,7 +30,8 @@ public class AssetsUtils {
             Log.e(TAG, fileName);
             AssetFileDescriptor openFd = context.getAssets().openFd(fileName);
 
-            if (force) {
+            if (force && destFile.exists()) {
+
                 destFile.delete();
             } else {
                 if (destFile.exists()) {
